server/requirements.txt
fastapi
uvicorn
pydantic
openenv-core
requests
openai